requirements.txt
fastapi==0.104.1
pydantic-settings==2.1.0
uvicorn==0.17.6
sqlalchemy==1.4.32
python-dotenv==0.21.0
validators==0.18.2
